Add error handlers for the cline-core process.
Move the ProtoBus service into its own file.
