+++ src/standalone/cline-core.ts
@@ -1,67 +1,27 @@
-import * as grpc from "@grpc/grpc-js"
-import { ReflectionService } from "@grpc/reflection"
-import * as health from "grpc-health-check"
-import * as hostProviders from "@hosts/host-providers"
 import { activate } from "@/extension"
 import { Controller } from "@core/controller"
-import { extensionContext, outputChannel, postMessage } from "./vscode-context"
-import { getPackageDefinition, log } from "./utils"
-import { GrpcHandler, GrpcStreamingResponseHandler } from "@hosts/external/grpc-types"
-import { addProtobusServices } from "@generated/standalone/server-setup"
-import { StreamingResponseHandler } from "@core/controller/grpc-handler"
-import { ExternalHostBridgeClientManager } from "@hosts/external/host-bridge-client-manager"
+import { ExternalDiffViewProvider } from "@hosts/external/ExternalDiffviewProvider"
 import { ExternalWebviewProvider } from "@hosts/external/ExternalWebviewProvider"
+import { ExternalHostBridgeClientManager } from "@hosts/external/host-bridge-client-manager"
+import * as hostProviders from "@hosts/host-providers"
 import { WebviewProviderType } from "@shared/webview/types"
 import { v4 as uuidv4 } from "uuid"
-import { ExternalDiffViewProvider } from "@hosts/external/ExternalDiffviewProvider"
+import { log } from "./utils"
+import { extensionContext, outputChannel, postMessage } from "./vscode-context"
 
 export const PROTOBUS_PORT = 26040
 export const HOSTBRIDGE_PORT = 26041
 
 async function main() {
-	log("Starting standalone service...")
+	log("Starting cline-core service...")
+
+	// Set up global error handlers to prevent process crashes
+	setupGlobalErrorHandlers()
 
 	hostProviders.initializeHostProviders(createWebview, createDiffView, new ExternalHostBridgeClientManager())
 	activate(extensionContext)
 	const controller = new Controller(extensionContext, outputChannel, postMessage, uuidv4())
 	startProtobusService(controller)
-}
-
-function startProtobusService(controller: Controller) {
-	const server = new grpc.Server()
-
-	// Set up health check.
-	const healthImpl = new health.HealthImplementation({ "": "SERVING" })
-	healthImpl.addToServer(server)
-
-	// Add all the handlers for the ProtoBus services to the server.
-	addProtobusServices(server, controller, wrapHandler, wrapStreamingResponseHandler)
-
-	// Create reflection service with protobus service names
-	const packageDefinition = getPackageDefinition()
-	const reflection = new ReflectionService(packageDefinition, {
-		services: getProtobusServiceNames(packageDefinition),
-	})
-	reflection.addToServer(server)
-
-	// Start the server.
-	const host = process.env.PROTOBUS_ADDRESS || `127.0.0.1:${PROTOBUS_PORT}`
-	server.bindAsync(host, grpc.ServerCredentials.createInsecure(), (err) => {
-		if (err) {
-			log(`Error: Failed to bind to ${host}, port may be unavailable. ${err.message}`)
-			process.exit(1)
-		}
-		server.start()
-		log(`gRPC server listening on ${host}`)
-	})
-}
-
-function getProtobusServiceNames(packageDefinition: { [x: string]: any }): string[] {
-	// Filter service names to only include cline services
-	const protobusServiceNames = Object.keys(packageDefinition).filter(
-		(name) => name.startsWith("cline.") || name.startsWith("grpc.health"),
-	)
-	return protobusServiceNames
 }
 
 function createWebview() {
@@ -72,66 +32,47 @@ function createDiffView() {
 }
 
 /**
- * Wraps a Promise-based handler function to make it compatible with gRPC's callback-based API.
- * This function converts an async handler that returns a Promise into a function that uses
- * the gRPC callback pattern.
- *
- * @template TRequest - The type of the request object
- * @template TResponse - The type of the response object
- * @param handler - The Promise-based handler function to wrap
- * @param controllerInstance - The controller instance to pass to the handler
- * @returns A gRPC-compatible callback-style handler function
+ * Sets up global error handlers to prevent the process from crashing
+ * on unhandled exceptions and promise rejections
  */
-function wrapHandler<TRequest, TResponse>(
-	handler: GrpcHandler<TRequest, TResponse>,
-	controller: Controller,
-): grpc.handleUnaryCall<TRequest, TResponse> {
-	return async (call: grpc.ServerUnaryCall<TRequest, TResponse>, callback: grpc.sendUnaryData<TResponse>) => {
-		try {
-			log(`gRPC request: ${call.getPath()}`)
-			const result = await handler(controller, call.request)
-			callback(null, result)
-		} catch (err: any) {
-			log(`gRPC handler error: ${call.getPath()}\n${err.stack}`)
-			callback({
-				code: grpc.status.INTERNAL,
-				message: err.message || "Internal error",
-			} as grpc.ServiceError)
+function setupGlobalErrorHandlers() {
+	// Handle unhandled exceptions
+	process.on("uncaughtException", (error: Error) => {
+		log(`Uncaught Exception: ${error.message}`)
+		log(`Stack trace: ${error.stack}`)
+		// Log the error but don't exit the process
+	})
+
+	// Handle unhandled promise rejections
+	process.on("unhandledRejection", (reason: any, promise: Promise<any>) => {
+		log(`Unhandled Promise Rejection at: ${promise}`)
+		log(`Reason: ${reason}`)
+		if (reason instanceof Error) {
+			log(`Stack trace: ${reason.stack}`)
 		}
-	}
-}
+		// Log the error but don't exit the process
+	})
 
-function wrapStreamingResponseHandler<TRequest, TResponse>(
-	handler: GrpcStreamingResponseHandler<TRequest, TResponse>,
-	controller: Controller,
-): grpc.handleServerStreamingCall<TRequest, TResponse> {
-	return async (call: grpc.ServerWritableStream<TRequest, TResponse>) => {
-		try {
-			const requestId = call.metadata.get("request-id").pop()?.toString()
-			log(`gRPC streaming request: ${call.getPath()}`)
-
-			const responseHandler: StreamingResponseHandler = (response, isLast, sequenceNumber) => {
-				try {
-					call.write(response) // Use a bound version of call.write to maintain proper 'this' context
-
-					if (isLast === true) {
-						log(`Closing stream for ${requestId}`)
-						call.end()
-					}
-					return Promise.resolve()
-				} catch (error) {
-					return Promise.reject(error)
-				}
-			}
-			await handler(controller, call.request, responseHandler, requestId)
-		} catch (err: any) {
-			log(`gRPC handler error: ${call.getPath()}\n${err.stack}`)
-			call.destroy({
-				code: grpc.status.INTERNAL,
-				message: err.message || "Internal error",
-			} as grpc.ServiceError)
+	// Handle process warnings (optional, for debugging)
+	process.on("warning", (warning: Error) => {
+		log(`Process Warning: ${warning.name}: ${warning.message}`)
+		if (warning.stack) {
+			log(`Stack trace: ${warning.stack}`)
 		}
-	}
+	})
+
+	// Graceful shutdown handlers
+	process.on("SIGINT", () => {
+		log("Received SIGINT, shutting down gracefully...")
+		process.exit(0)
+	})
+
+	process.on("SIGTERM", () => {
+		log("Received SIGTERM, shutting down gracefully...")
+		process.exit(0)
+	})
+
+	log("Global error handlers set up successfully")
 }
 
 main()
